refactor: improve package organization
