+++ packages/core/package.json
@@ -5,11 +5,15 @@
   "description": "Ellmers is a tool for building and running DAG pipelines of AI tasks.",
   "scripts": {
     "watch": "concurrently -c 'auto' 'bun:watch-*'",
-    "watch-code": "bun build --watch --no-clear-screen --target=browser --sourcemap=external --external @huggingface/transformers --outdir ./dist ./src/index.ts",
+    "watch-browser": "bun build --watch --no-clear-screen --target=browser --sourcemap=external --outdir ./dist ./src/index.ts",
+    "no-watch-node": "bun build --watch --no-clear-screen --target=node --sourcemap=external --outdir ./dist ./src/index.ts --entry-naming node.js",
+    "no-watch-bun": "bun build --watch --no-clear-screen --target=bun --sourcemap=external --outdir ./dist ./src/index.ts --entry-naming bun.js",
     "watch-types": "tsc --watch --preserveWatchOutput",
-    "build": "bun run build-clean && bun run build-types && bun run build-code",
+    "build": "bun run build-clean && bun run build-types && bun run build-browser",
     "build-clean": "rm -fr dist/* tsconfig.tsbuildinfo",
-    "build-code": "bun build --target=browser --sourcemap=external --external @huggingface/transformers --outdir ./dist ./src/index.ts",
+    "build-browser": "bun build --target=browser --sourcemap=external --outdir ./dist ./src/index.ts",
+    "build-node": "bun build --target=node --sourcemap=external --outdir ./dist ./src/index.ts --entry-naming node.js",
+    "build-bun": "bun build --target=bun --sourcemap=external --outdir ./dist ./src/index.ts --entry-naming bun.js",
     "build-types": "tsc",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "test": "bun test"
