+++ packages/core/src/index.ts
@@ -9,9 +9,6 @@ export * from "./source/Document";
 export * from "./source/DocumentConverterText";
 export * from "./source/DocumentConverterMarkdown";
 export * from "./task/index";
-export * from "./provider/ProviderRegistry";
-export * from "./model/Model";
-export * from "./model/InMemoryStorage";
 export * from "./storage/base/KVRepository";
 export * from "./storage/taskoutput/TaskOutputRepository";
 export * from "./storage/taskoutput/InMemoryTaskOutputRepository";
