+++ packages/core/src/util/Misc.ts
@@ -60,13 +60,10 @@ export function deepEqual(a: any, b: any): boolean {
 export function sortObject(obj: Record<string, any>): Record<string, any> {
   return Object.keys(obj)
     .sort()
-    .reduce(
-      (result, key) => {
-        result[key] = obj[key];
-        return result;
-      },
-      {} as Record<string, any>
-    );
+    .reduce((result, key) => {
+      result[key] = obj[key];
+      return result;
+    }, {} as Record<string, any>);
 }
 
 export function serialize(obj: Record<string, any>): string {
@@ -85,8 +82,7 @@ export async function sha256(data: string): Promise<string> {
   } else if (typeof process === "object" && process.versions && process.versions.node) {
     // Node.js environment
     // eslint-disable-next-line @typescript-eslint/no-var-requires
-    const crypto = require("crypto");
-    return Promise.resolve(crypto.createHash("sha256").update(data).digest("hex"));
+    return Promise.resolve(new Bun.CryptoHasher("sha256").update(data).digest("hex"));
   } else {
     throw new Error("Unsupported environment");
   }
