+++ packages/provider/package.json
@@ -5,23 +5,23 @@
   "description": "Ellmers is a tool for building and running DAG pipelines of AI tasks.",
   "scripts": {
     "watch": "concurrently -c 'auto' 'bun:watch-*'",
-    "watch-hf-transformers-browser": "bun build --watch --no-clear-screen --target=browser --sourcemap=external --external @huggingface/transformers --external ellmers-core --outdir ./dist/hf-transformers ./src/hf-transformers/browser.ts",
-    "watch-hf-transformers-server": "bun build --watch --no-clear-screen --target=node --sourcemap=external --external @huggingface/transformers --external ellmers-core --outdir ./dist/hf-transformers ./src/hf-transformers/server.ts",
-    "watch-hf-transformers-worker": "bun build --watch --no-clear-screen --target=bun --sourcemap=external --external @huggingface/transformers --external ellmers-core --outdir ./dist/hf-transformers ./src/hf-transformers/worker.ts",
-    "watch-tf-mediapipe-browser": "bun build --watch --no-clear-screen --target=browser --sourcemap=external --external @mediapipe/tasks-text --external ellmers-core --outdir ./dist/tf-mediapipe ./src/tf-mediapipe/browser.ts",
-    "watch-tf-mediapipe-server": "bun build --watch --no-clear-screen --target=node --sourcemap=external --external @mediapipe/tasks-text --external ellmers-core --outdir ./dist/tf-mediapipe ./src/tf-mediapipe/server.ts",
-    "watch-tf-mediapipe-worker": "bun build --watch --no-clear-screen --target=bun --sourcemap=external --external @mediapipe/tasks-text --external ellmers-core --outdir ./dist/tf-mediapipe ./src/tf-mediapipe/worker.ts",
+    "watch-hf-transformers-browser": "bun build --watch --no-clear-screen --target=browser --sourcemap=external --external @huggingface/transformers --external ellmers-core --external ellmers-task-llm --outdir ./dist/hf-transformers ./src/hf-transformers/browser.ts",
+    "watch-hf-transformers-server": "bun build --watch --no-clear-screen --target=node --sourcemap=external --external @huggingface/transformers --external ellmers-core --external ellmers-task-llm --outdir ./dist/hf-transformers ./src/hf-transformers/server.ts",
+    "watch-hf-transformers-worker": "bun build --watch --no-clear-screen --target=bun --sourcemap=external --external @huggingface/transformers --external ellmers-core --external ellmers-task-llm --outdir ./dist/hf-transformers ./src/hf-transformers/worker.ts",
+    "watch-tf-mediapipe-browser": "bun build --watch --no-clear-screen --target=browser --sourcemap=external --external @mediapipe/tasks-text --external ellmers-core --external ellmers-task-llm --outdir ./dist/tf-mediapipe ./src/tf-mediapipe/browser.ts",
+    "watch-tf-mediapipe-server": "bun build --watch --no-clear-screen --target=node --sourcemap=external --external @mediapipe/tasks-text --external ellmers-core --external ellmers-task-llm --outdir ./dist/tf-mediapipe ./src/tf-mediapipe/server.ts",
+    "watch-tf-mediapipe-worker": "bun build --watch --no-clear-screen --target=bun --sourcemap=external --external @mediapipe/tasks-text --external ellmers-core --external ellmers-task-llm --outdir ./dist/tf-mediapipe ./src/tf-mediapipe/worker.ts",
     "watch-types": "tsc --watch --preserveWatchOutput",
     "build": "bun run build-clean && bun run build-types && bun run build-hf-transformers && bun run build-tf-mediapipe",
     "build-clean": "rm -fr dist/* tsconfig.tsbuildinfo",
     "build-hf-transformers": "bun run build-hf-transformers-browser && bun run build-hf-transformers-server && bun run build-hf-transformers-worker",
-    "build-hf-transformers-browser": "bun build --target=browser --sourcemap=external --external @huggingface/transformers --external ellmers-core --outdir ./dist/hf-transformers ./src/hf-transformers/browser.ts",
-    "build-hf-transformers-server": "bun build --target=node --sourcemap=external --external @huggingface/transformers --external ellmers-core --outdir ./dist/hf-transformers ./src/hf-transformers/server.ts",
-    "build-hf-transformers-worker": "bun build --target=bun --sourcemap=external --external @huggingface/transformers --external ellmers-core --outdir ./dist/hf-transformers ./src/hf-transformers/worker.ts",
+    "build-hf-transformers-browser": "bun build --target=browser --sourcemap=external --external @huggingface/transformers --external ellmers-core --external ellmers-task-llm --outdir ./dist/hf-transformers ./src/hf-transformers/browser.ts",
+    "build-hf-transformers-server": "bun build --target=node --sourcemap=external --external @huggingface/transformers --external ellmers-core --external ellmers-task-llm --outdir ./dist/hf-transformers ./src/hf-transformers/server.ts",
+    "build-hf-transformers-worker": "bun build --target=bun --sourcemap=external --external @huggingface/transformers --external ellmers-core --external ellmers-task-llm --outdir ./dist/hf-transformers ./src/hf-transformers/worker.ts",
     "build-tf-mediapipe": "bun run build-tf-mediapipe-browser && bun run build-tf-mediapipe-server && bun run build-tf-mediapipe-worker",
-    "build-tf-mediapipe-browser": "bun build --target=browser --sourcemap=external --external @mediapipe/tasks-text --external ellmers-core --outdir ./dist/tf-mediapipe ./src/tf-mediapipe/browser.ts",
-    "build-tf-mediapipe-server": "bun build --target=node --sourcemap=external --external @mediapipe/tasks-text --external ellmers-core --outdir ./dist/tf-mediapipe ./src/tf-mediapipe/server.ts",
-    "build-tf-mediapipe-worker": "bun build --target=bun --sourcemap=external --external @mediapipe/tasks-text --external ellmers-core --outdir ./dist/tf-mediapipe ./src/tf-mediapipe/worker.ts",
+    "build-tf-mediapipe-browser": "bun build --target=browser --sourcemap=external --external @mediapipe/tasks-text --external ellmers-core --external ellmers-task-llm --outdir ./dist/tf-mediapipe ./src/tf-mediapipe/browser.ts",
+    "build-tf-mediapipe-server": "bun build --target=node --sourcemap=external --external @mediapipe/tasks-text --external ellmers-core --external ellmers-task-llm --outdir ./dist/tf-mediapipe ./src/tf-mediapipe/server.ts",
+    "build-tf-mediapipe-worker": "bun build --target=bun --sourcemap=external --external @mediapipe/tasks-text --external ellmers-core --external ellmers-task-llm --outdir ./dist/tf-mediapipe ./src/tf-mediapipe/worker.ts",
     "build-types": "tsc",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "test": "bun test"
