+++ packages/provider/src/ggml-local/model/GgmlLocalModel.ts
@@ -5,4 +5,16 @@
 //    *   Licensed under the Apache License, Version 2.0 (the "License");           *
 //    *******************************************************************************
 
-export * from "./Model";
+import {
+  Model,
+  ModelOptions,
+  ModelProcessorEnum,
+  ModelUseCaseEnum,
+} from "../../../../task-llm/src/model/Model";
+
+export class GgmlLocalModel extends Model {
+  constructor(name: string, useCase: ModelUseCaseEnum[], options?: ModelOptions) {
+    super(name, useCase, options);
+  }
+  readonly type = ModelProcessorEnum.LOCAL_LLAMACPP;
+}
