+++ packages/provider/src/hf-transformers/bindings/all_inmemory.ts
@@ -1,11 +1,6 @@
-import {
-  getProviderRegistry,
-  InMemoryJobQueue,
-  ModelProcessorEnum,
-  ConcurrencyLimiter,
-  TaskInput,
-  TaskOutput,
-} from "ellmers-core";
+import { InMemoryJobQueue, ConcurrencyLimiter, TaskInput, TaskOutput } from "ellmers-core";
+import { getProviderRegistry, ModelProcessorEnum } from "ellmers-task-llm";
+
 import { registerHuggingfaceLocalTasks } from "./local_hf";
 import "../model/ONNXModelSamples";
