+++ packages/provider/src/tf-mediapipe/model/MediaPipeModel.ts
@@ -5,7 +5,7 @@
 //    *   Licensed under the Apache License, Version 2.0 (the "License");           *
 //    *******************************************************************************
 
-import { Model, ModelOptions, ModelProcessorEnum, ModelUseCaseEnum } from "ellmers-core";
+import { Model, ModelOptions, ModelProcessorEnum, ModelUseCaseEnum } from "ellmers-task-llm";
 
 export class MediaPipeTfJsModel extends Model {
   constructor(
