+++ packages/task-llm/src/index.ts
@@ -1,1 +1,4 @@
 export * from "./task";
+export * from "./model/Model";
+export * from "./model/InMemoryStorage";
+export * from "./provider/ProviderRegistry";
