+++ samples/cli/package.json
@@ -5,11 +5,11 @@
   "description": "Ellmers is a tool for building and running DAG pipelines of AI tasks.",
   "scripts": {
     "watch": "concurrently -c 'auto' -n 'cli:' 'bun:watch-*'",
-    "watch-js": "bun build --watch --target=node --sourcemap=external --external listr2 --external @huggingface/transformers --outdir ./dist ./src/lib.ts ./src/ellmers.ts",
+    "watch-js": "bun build --watch --target=node --sourcemap=external --splitting --outdir ./dist ./src/lib.ts ./src/ellmers.ts",
     "watch-types": "tsc --watch --preserveWatchOutput",
     "build": "bun run build-clean && bun run build-types && bun run build-js",
     "build-clean": "rm -fr dist/* tsconfig.tsbuildinfo",
-    "build-js": "bun build --target=node --sourcemap=external --external listr2 --external @huggingface/transformers --outdir ./dist ./src/lib.ts ./src/ellmers.ts",
+    "build-js": "bun build --target=node --sourcemap=external --splitting --outdir ./dist ./src/lib.ts ./src/ellmers.ts",
     "build-types": "tsc",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
