+++ samples/web/src/RunGraphFlow.tsx
@@ -1,4 +1,4 @@
-import React, { Dispatch, SetStateAction, useCallback, useEffect, useRef } from "react";
+import React, { Dispatch, SetStateAction, useEffect, useRef } from "react";
 import {
   ReactFlow,
   Controls,
